Former-commit-id: 8d7f0d1293e06082b54941e435a8a9e4c56579d3 [formerly 8c72791f1145b1c9dc4a17bb8e60baf9bffa99d7]
Former-commit-id: eeb10df93d8aa5b9d36f9cb345bca76741c5b315

+++ app/Http/Controllers/ImportationController.php
@@ -1420,11 +1420,15 @@ class ImportationController extends Controller
         }
 
         $failrate = FailRate::find($id);
-        $failrate->forceDelete();
-        $request->session()->flash('message.content', 'Updated Rate' );
-        $request->session()->flash('message.nivel', 'success');
-        $request->session()->flash('message.title', 'Well done!');
-        return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
+            if(is_null($failrate)){
+                return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
+        }else{
+            $failrate->forceDelete();
+            $request->session()->flash('message.content', 'Updated Rate' );
+            $request->session()->flash('message.nivel', 'success');
+            $request->session()->flash('message.title', 'Well done!');
+            return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
+        }
     }
     public function UpdateRatesD(Request $request, $id){
         //dd($request->all());
@@ -2020,12 +2024,15 @@ class ImportationController extends Controller
                 ]);
             }
         }
-        $failSurcharge->forceDelete();
-        $request->session()->flash('message.content', 'Surcharge Updated' );
-        $request->session()->flash('message.nivel', 'success');
-        $request->session()->flash('message.title', 'Well done!');
-        return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
-
+        if(is_null($failSurcharge)){
+            return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
+        }else{
+            $failSurcharge->forceDelete();
+            $request->session()->flash('message.content', 'Surcharge Updated' );
+            $request->session()->flash('message.nivel', 'success');
+            $request->session()->flash('message.title', 'Well done!');
+            return redirect()->route('Failed.Developer.For.Contracts',[$request->contract_id,$request->nameTab]);
+        }
     }
 
     public function UpdateSurchargersD(Request $request, $id){
